### 2023-07-06 #5625 @editorialbot
[REVIEW]: Lightning-UDA-Detect: Easily run unsupervised domain adaptation object detection
**Submitting author:** @just-eoghan (<a href="http://orcid.org/0000-0001-7394-1926">Eoghan Mulcahy</a>)
**Repository:** https://github.com/just-eoghan/Lightning-UDA-detect
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
