### 2023-07-19 #5625 @ns-rse
Hi @just-eoghan,

Sorry for leaving this half-way through, I've tidied up the above checklist, removed the comments and include them below under each section.

As to when to address them I'm not sure, this is my first time reviewing for JOSS. It might perhaps be worth waiting for feedback from other reviewers.

@ns-rse

## General checks

###  **Repository** : Is the source code for this software available at the repository url?

**Yes** Source code is available at the URL <https://github.com/just-eoghan/Lightning-UDA-detect/>

###  **Contribution and authorship** : Has the submitting author made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

The submitting author has made major contributions to the software, with all `git blame` checked attributed to
`just-eoghan`. There are no commits from other authors and I assume the co-authors are supervisors of the work.

## Functionality

### **Installation**: Does installation proceed as outlined in the documentation?

**NO**  - It might be an obvious step to experienced users but this package is not available on PyPI nor Conda Forge. This
means that users have to clone the repository from Git and `cd` into it after creating their Conda environment...

    git clone https://github.com/just-eoghan/Lightning-UDA-detect.git
    cd Lightning-UDA-detect

...only then can they `pip install -r requirements.txt`. The package would benefit from having the install requirements
documented in `pyproject.toml` (the current recommended configuration file for installation with
[Setuptools](https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html) framework). This would also allow the
inclusion of metadata about the package, configuration of the `black` linter which I see is mentioned in some of the
dependencies (and installation of such dependencies can be grouped under `project.optional-dependencies`), this helps set the stage for publishing the package on PyPI.

### **Functionality**: Have the functional claims of the software been confirmed?

**Maybe** - The claim is that the package simplifies performing certain analyses which otherwise required abandoned
packages that have more convoluted installation processes. These have not been attempted but it does appear the process
is easier compared to the instructions that have been read although this reviewer was unable to complete running the
example.

###  **Performance**: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
Its claimed the precision of the models used here exceeds the original but I have not checked this.

## Documentation

###  **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

**Yes** authors make it clear that existing tools
([maskrcnn-benchmark](https://github.com/facebookresearch/maskrcnn-benchmark)) is convoluted and having read the
installation instructions for this tool (which hasn't been updated in 4 years) the presented software
`Lightning-UDA-detect` does seem considerably to be easier to install and use.

###  **Installation instructions**: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

**Yes** dependencies are detailed in a `requirements.txt` and installation instructions when followed (subject to the above
comment about needing to `git clone` the repository) result in successful installation and an environment that allows
the code to run. As noted though, there is no adherence to any of the Python packaging frameworks such as Setuptools,
Poetry or PDM.

Perhaps a more explicit statement of the Python versions supported would be useful (e.g. badges). Installation of
dependencies using Python 3.11 failed with

```
    Collecting notebook-shim>=0.2.3 (from nbclassic->jupyterlab==3.6.3->-r requirements.txt (line 28))
      Using cached notebook_shim-0.2.3-py3-none-any.whl (13 kB)
    INFO: pip is looking at multiple versions of opencv-python-headless to determine which version is compatible with other requirements. This could take a while.
    Collecting opencv-python-headless>=4.1.1 (from albumentations==1.3.0->-r requirements.txt (line 30))
      Downloading opencv_python_headless-4.7.0.72-cp37-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (49.2 MB)
      Downloading opencv_python_headless-4.7.0.68-cp37-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (49.2 MB)
         ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 49.2/49.2 MB 763.5 kB/s eta 0:00:00
      Downloading opencv_python_headless-4.6.0.66-cp36-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (48.3 MB)
         ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 48.3/48.3 MB 564.1 kB/s eta 0:00:00
      Downloading opencv_python_headless-4.5.5.64-cp36-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (47.8 MB)
         ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 47.8/47.8 MB 650.2 kB/s eta 0:00:00
      Downloading opencv_python_headless-4.5.5.62-cp36-abi3-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (47.7 MB)
         ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 47.7/47.7 MB 856.2 kB/s eta 0:00:00
      Downloading opencv-python-headless-4.5.4.60.tar.gz (89.8 MB)
         ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 89.8/89.8 MB 553.7 kB/s eta 0:00:00
      Installing build dependencies ... error
      error: subprocess-exited-with-error

      × pip subprocess to install build dependencies did not run successfully.
      │ exit code: 1
      ╰─> [20 lines of output]
          Ignoring numpy: markers 'python_version == "3.6" and platform_machine != "aarch64" and platform_machine != "arm64"' don't match your environment
          Ignoring numpy: markers 'python_version == "3.7" and platform_machine != "aarch64" and platform_machine != "arm64"' don't match your environment
          Ignoring numpy: markers 'python_version == "3.8" and platform_machine != "aarch64" and platform_machine != "arm64"' don't match your environment
          Ignoring numpy: markers 'python_version <= "3.9" and sys_platform == "linux" and platform_machine == "aarch64"' don't match your environment
          Ignoring numpy: markers 'python_version <= "3.9" and sys_platform == "darwin" and platform_machine == "arm64"' don't match your environment
          Ignoring numpy: markers 'python_version == "3.9" and platform_machine != "aarch64" and platform_machine != "arm64"' don't match your environment
          Collecting setuptools
            Using cached setuptools-68.0.0-py3-none-any.whl (804 kB)
          Collecting wheel
            Using cached wheel-0.40.0-py3-none-any.whl (64 kB)
          Collecting scikit-build
            Downloading scikit_build-0.17.6-py3-none-any.whl (84 kB)
               ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 84.3/84.3 kB 566.3 kB/s eta 0:00:00
          Collecting cmake
            Using cached cmake-3.26.4-py2.py3-none-manylinux2014_x86_64.manylinux_2_17_x86_64.whl (24.0 MB)
          Collecting pip
            Using cached pip-23.1.2-py3-none-any.whl (2.1 MB)
          ERROR: Ignored the following versions that require a different python version: 1.21.2 Requires-Python >=3.7,<3.11; 1.21.3 Requires-Python >=3.7,<3.11; 1.21.4 Requires-Python >=3.7,<3.11; 1.21.5 Requires-Python >=3.7,<3.11; 1.21.6 Requires-Python >=3.7,<3.11
          ERROR: Could not find a version that satisfies the requirement numpy==1.21.2 (from versions: 1.3.0, 1.4.1, 1.5.0, 1.5.1, 1.6.0, 1.6.1, 1.6.2, 1.7.0, 1.7.1, 1.7.2, 1.8.0, 1.8.1, 1.8.2, 1.9.0, 1.9.1, 1.9.2, 1.9.3, 1.10.0.post2, 1.10.1, 1.10.2, 1.10.4, 1.11.0, 1.11.1, 1.11.2, 1.11.3, 1.12.0, 1.12.1, 1.13.0, 1.13.1, 1.13.3, 1.14.0, 1.14.1, 1.14.2, 1.14.3, 1.14.4, 1.14.5, 1.14.6, 1.15.0, 1.15.1, 1.15.2, 1.15.3, 1.15.4, 1.16.0, 1.16.1, 1.16.2, 1.16.3, 1.16.4, 1.16.5, 1.16.6, 1.17.0, 1.17.1, 1.17.2, 1.17.3, 1.17.4, 1.17.5, 1.18.0, 1.18.1, 1.18.2, 1.18.3, 1.18.4, 1.18.5, 1.19.0, 1.19.1, 1.19.2, 1.19.3, 1.19.4, 1.19.5, 1.20.0, 1.20.1, 1.20.2, 1.20.3, 1.21.0, 1.21.1, 1.22.0, 1.22.1, 1.22.2, 1.22.3, 1.22.4, 1.23.0rc1, 1.23.0rc2, 1.23.0rc3, 1.23.0, 1.23.1, 1.23.2, 1.23.3, 1.23.4, 1.23.5, 1.24.0rc1, 1.24.0rc2, 1.24.0, 1.24.1, 1.24.2, 1.24.3, 1.24.4, 1.25.0rc1, 1.25.0)
          ERROR: No matching distribution found for numpy==1.21.2
          [end of output]

    error: subprocess-exited-with-error

    × pip subprocess to install build dependencies did not run successfully.
    │ exit code: 1
    ╰─> See above for output.

```

###  **Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

**Yes** - An example is provided using data that is freely available. Having gone through the example the following
observations were made.

-   Requires the user to register/signin with the data provider this could be mentioned in the instructions.
-   The instructions section on **Weights and Biases** states&#x2026;

> **Weights and Biases**
>
> You can automatically log runs to your W&B account by editing configs/logger/wandb
>
> Set entity to the name of your W&B account. Set project to the name of the project you have created in W&B.

This reviewer is not familiar with *W&B*, but reading the noted config file (which is actually
`configs/logger/wandb.yaml`) indicates its the site <https://wandb.ai> better documentation explaining what this is and
how it works would be beneficial.

-   Running the example results in additional data being downloaded this could be mentioned up-front (although at ~98Mb
    in comparison to what has already been downloaded this is negligible).
-   The first command using local logging failed...

```
    ❱ python run.py experiment=mic_da_cityscapes logger=csv
    [2023-07-09 18:16:35,720][torch.distributed.nn.jit.instantiator][INFO] - Created a temporary directory at /tmp/tmpeth8dh8m
    [2023-07-09 18:16:35,721][torch.distributed.nn.jit.instantiator][INFO] - Writing /tmp/tmpeth8dh8m/_remote_module_non_scriptable.py
    [2023-07-09 18:16:35,789][src.utils.utils][INFO] - Disabling python warnings! <config.ignore_warnings=True>
    │   └── _target_: pytorch_lightning.Trainer
    ├── trainer
    [2023-07-09 18:20:52,445][src.train][INFO] - Instantiating trainer <pytorch_lightning.Trainer>
    │       min_epochs: 1
    │       max_epochs: -1
    │       deterministic: false
    │       num_sanity_val_steps: 0
    │       min_steps: 1
    │       devices:
    │       accelerator: gpu
        self._accelerator_connector = _AcceleratorConnector(
    │       - 0
    │
    ├── model
    │   └── _target_: src.models.mic_da_frcnn.MicDaFrcnnDetectionModel
    │       da_net:
    │       weight_decay: 0.0001
    lightning_fabric.utilities.exceptions.MisconfigurationException: No supported gpu backend found!
    │         _target_: src.models.modules.da_mic.model.MicDaFRCNN
    │         backbone_out_channels: 256
    │         consit_weight: 0.1
    │         img_grl_weight: 0.01
    │         ins_grl_weight: 0.1
    │         num_classes: 9
    │       lr_scheduler: true
    │       num_classes: 9
    │
    ├── datamodule
    │   └── _target_: src.datamodules.da_cityscapes_datamodule.DaCityscapesDatamodule
    │       source_data_dir: /mnt/work/git/hub/Lightning-UDA-detect/data/cityscapes/
    │       target_data_dir: /mnt/work/git/hub/Lightning-UDA-detect/data/foggy_cityscapes/
    │       train_batch_size: 2
    │       pin_memory: false
    │       num_workers: 0
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/pytorch_lightning/utilities/argparse.py", line 69, in insert_env_defaults
    │       num_classes: 9
    │       image_width: 1600
    │       image_height: 800
    │
    ├── callbacks
    │   └── model_checkpoint:
    │         _target_: pytorch_lightning.callbacks.ModelCheckpoint
    │         monitor: val/map50
    │         mode: max
    │         save_top_k: 1
    │         save_last: true
    │         verbose: false
    │         auto_insert_metric_name: false
    │       model_summary:
    (joss-uda-detect) ❱
    │       model_summary:
    │         _target_: pytorch_lightning.callbacks.RichModelSummary
    │         max_depth: 4
    │       rich_progress_bar:
    │         _target_: pytorch_lightning.callbacks.RichProgressBar
    │
    ├── logger
    │   └── csv:
    │         _target_: pytorch_lightning.loggers.csv_logs.CSVLogger
    │         save_dir: .
    │         name: csv/
    │         version: mic_da_cityscapes
    │         prefix: ''
    │
    ├── test_after_training
    │   └── False
    ├── seed
    │   └── 100101101112115101101107046097105
    └── name
        └── mic_da_cityscapes
    [2023-07-09 18:16:35,848][lightning_fabric.utilities.seed][INFO] - Global seed set to 1292028720
    [2023-07-09 18:16:35,850][src.train][INFO] - Instantiating datamodule <src.datamodules.da_cityscapes_datamodule.DaCityscapesDatamodule>
    [2023-07-09 18:16:36,750][src.train][INFO] - Instantiating model <src.models.mic_da_frcnn.MicDaFrcnnDetectionModel>
    Downloading: "https://download.pytorch.org/models/resnet50-11ad3fa6.pth" to /home/neil/.cache/torch/hub/checkpoints/resnet50-11ad3fa6.pth
    100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 97.8M/97.8M [04:14<00:00, 403kB/s]
    [Masking] Use color augmentation.
    [2023-07-09 18:20:52,440][src.train][INFO] - Instantiating callback <pytorch_lightning.callbacks.ModelCheckpoint>
    [2023-07-09 18:20:52,442][src.train][INFO] - Instantiating callback <pytorch_lightning.callbacks.RichModelSummary>
    [2023-07-09 18:20:52,443][src.train][INFO] - Instantiating callback <pytorch_lightning.callbacks.RichProgressBar>
    [2023-07-09 18:20:52,444][src.train][INFO] - Instantiating logger <pytorch_lightning.loggers.csv_logs.CSVLogger>
    Traceback (most recent call last):
    Error executing job with overrides: ['experiment=mic_da_cityscapes', 'logger=csv']
    Traceback (most recent call last):
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/hydra/_internal/instantiate/_instantiate2.py", line 62, in _call_target
        return target(*args, **kwargs)
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/pytorch_lightning/utilities/argparse.py", line 69, in insert_env_defaults
        return fn(self, **kwargs)
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/pytorch_lightning/trainer/trainer.py", line 393, in __init__
    __
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/pytorch_lightning/trainer/connectors/accelerator_connector.py", line 154, in __init__
        self._accelerator_flag = self._choose_gpu_accelerator_backend()
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/pytorch_lightning/trainer/connectors/accelerator_connector.py", line 371, in _choose_gpu_accelerator_backend
        raise MisconfigurationException("No supported gpu backend found!")

    Traceback (most recent call last):
      File "run.py", line 30, in main
        return train(config)
      File "/mnt/work/git/hub/Lightning-UDA-detect/src/train.py", line 61, in train
        trainer: Trainer = hydra.utils.instantiate(
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/hydra/_internal/instantiate/_instantiate2.py", line 180, in instantiate
        return instantiate_node(config, *args, recursive=_recursive_, convert=_convert_)
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/hydra/_internal/instantiate/_instantiate2.py", line 249, in instantiate_node
        return _call_target(target, *args, **kwargs)
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/hydra/_internal/instantiate/_instantiate2.py", line 64, in _call_target
        raise type(e)(
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/hydra/_internal/instantiate/_instantiate2.py", line 62, in _call_target
        return target(*args, **kwargs)
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/pytorch_lightning/trainer/trainer.py", line 393, in __init__
        return fn(self, **kwargs)
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/pytorch_lightning/trainer/trainer.py", line 393, in __init__
        self._accelerator_connector = _AcceleratorConnector(
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/pytorch_lightning/trainer/connectors/accelerator_connector.py", line 154, in __init__
        self._accelerator_flag = self._choose_gpu_accelerator_backend()
      File "/home/neil/.virtualenvs/joss-uda-detect/lib/python3.8/site-packages/pytorch_lightning/trainer/connectors/accelerator_connector.py", line 371, in _choose_gpu_accelerator_backend
        raise MisconfigurationException("No supported gpu backend found!")
    lightning_fabric.utilities.exceptions.MisconfigurationException: Error instantiating 'pytorch_lightning.trainer.trainer.Trainer' : No supported gpu backend found!

    Set the environment variable HYDRA_FULL_ERROR=1 for a complete stack trace.

...which appears to be because I do not have a dedicated Graphics card to provision GPU on the computer I'm trying
on. It would therefore be useful to state up-front that the package is contingent on having dedicated GPU hardware. I
 switched to a laptop with a dedicated GPU with ~4GB RAM, I then encountered memory errors

     File "/home/neil/miniconda3/envs/joss-lightning/lib/python3.8/site-packages/torch/autograd/__init__.py", line 200, in backward
        Variable._execution_engine.run_backward(  # Calls into the C++ engine to run the backward pass
    torch.cuda.OutOfMemoryError: CUDA out of memory. Tried to allocate 222.00 MiB (GPU 0; 3.81 GiB total capacity; 2.61 GiB already allocated; 72.69 MiB free; 3.63 GiB reserved in total by PyTorch) If reserved memory is >> allocated memory try setting max_split_size_mb to avoid fragmentation.  See documentation for Memory Management and PYTORCH_CUDA_ALLOC_CONF

    Set the environment variable HYDRA_FULL_ERROR=1 for a complete stack trace.
```
I switched a system with higher specification (GM200 TITAN X) and the job ran to completion. It might therefore be prudent to make it clearer up-front what the hardware requirements are.

-   The instructions say that parameters can be modified by editing files under `config/experiments` but no guidance is given as to what parameters might be tuned and the implications of doing so. Each file references additional configurations that are located under `configs/model/*.yaml` and there is no reference to these in the instructions which would likely be prudent.
-   Perhaps more importantly there is no indication of how the examples generalise to potential users datasets in terms of what data is required for training, what objects are detected and classified nor what parameters might be considered reasonable when doing so.

###  **Functionality documentation**: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

**Maybe** There is no API as such, rather a set of scripts that run using specified scripts. As mentioned above there is no description of how parameters in the model might be adjusted and the impact this might have on the results, nor how one might go about deriving parameters for new data.

###  **Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?

**Missing** - no automated test-suite is included.

###  **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

[…]
