### 2023-08-04 #5625 @samirmartins
Dear colleagues (@galessiorob, @just-eoghan),

Here you can find a portion of my review. I will complete it once CityScapes approves my registration request (as seen below, this is an important issue that should be addressed by authors). If you have any questions, feel free to contact me.

Best regards,

-------

- I suggest that the authors create a pip installer (PyPI) to make the process more user-friendly. As it stands, everything is "script-based," and the final user has to install some unnecessary libraries (such as black) to run the scripts."

- I encountered an error during the installation of requirements.txt on Windows 11 (resolved by installing pycocotools-windows instead of the pycocotools from requirements). How can we resolve this OS-dependent issue? The error description can be found at the end of this message.

- To conduct the review, I need to wait for a few days until CityScapes admin approves my request to download the data and test the software. Is it possible to use any open-source data source for testing the tool and include it in the repository? This has, for instance, hindered my ability to test the code while first conducting the article review.

- According to the JOSS Reviewer Guidelines, the article should have documentation, tests, continuous integration, and a license. Some highlights: i) I did not find any tests in the repository; ii) The documentation is very concise and should be improved (currently, it is limited to README.md). As a suggestion, consider using mkdocs or a similar tool to build it; iii) the software has a MIT licence and; iv) allows continuous integration.

- As pointed out by @ns-rse, some references are missing DOIs.

- All commits were made solely by @just-eoghan. Where are the contributions from the other authors of the work?

- There are no clear guidelines on how to contribute to the software/report issues/seek support.

--------------------- MENTIONED INSTALLATION ERROR ---------------------

Building wheels for collected packages: pudb, pycocotools, antlr4-python3-runtime, urwid, pathtools, urwid_readline, pyperclip
  Building wheel for pudb (setup.py) ... done
  Created wheel for pudb: filename=pudb-2022.1.3-py3-none-any.whl size=70245 sha256=a3f61a016a36c7fa3e7f2f85a62a3011decccee8dd5691493a409c1022089751
  Stored in directory: c:\users\samir\appdata\local\pip\cache\wheels\5c\38\91\a0a7fa58d205af466e4e732c9d8ca00ee2cdb2e675a03811bc
  Building wheel for pycocotools (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for pycocotools (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [17 lines of output]
      running bdist_wheel
      running build
      running build_py
      creating build
      creating build\lib.win-amd64-cpython-38
      creating build\lib.win-amd64-cpython-38\pycocotools
      copying pycocotools\coco.py -> build\lib.win-amd64-cpython-38\pycocotools
      copying pycocotools\cocoeval.py -> build\lib.win-amd64-cpython-38\pycocotools
      copying pycocotools\mask.py -> build\lib.win-amd64-cpython-38\pycocotools
      copying pycocotools\__init__.py -> build\lib.win-amd64-cpython-38\pycocotools
      running build_ext
      Compiling pycocotools/_mask.pyx because it changed.
      [1/1] Cythonizing pycocotools/_mask.pyx
      building 'pycocotools._mask' extension
      C:\Users\Samir\AppData\Local\Temp\pip-build-env-6i4qqvl_\overlay\Lib\site-packages\Cython\Compiler\Main.py:381: FutureWarning: Cython directive 'language_level' not set, using '3str' for now (Py3). This has changed from earlier releases! File: C:\Users\Samir\AppData\Local\Temp\pip-install-1in0fdct\pycocotools_66ddde427fb54904a297c7c0f08b89f2\pycocotools\_mask.pyx
        tree = Parsing.p_module(s, pxd, full_module_name)
      error: Microsoft Visual C++ 14.0 or greater is required. Get it with "Microsoft C++ Build Tools": https://visualstudio.microsoft.com/visual-cpp-build-tools/
      [end of output]

  ERROR: Failed building wheel for pycocotools
  Building wheel for antlr4-python3-runtime (setup.py) ... done
  Created wheel for antlr4-python3-runtime: filename=antlr4_python3_runtime-4.8-py3-none-any.whl size=141217 sha256=534fb3f8502d16759b177ebaa4c6dfd47ab3eeb17579b68929fa084de75b4fc8
  Stored in directory: c:\users\samir\appdata\local\pip\cache\wheels\c8\d0\ab\d43c02eaddc5b9004db86950802442ad9a26f279c619e28da0
  Building wheel for urwid (setup.py) ... done
  Created wheel for urwid: filename=urwid-2.1.2-py3-none-any.whl size=235065 sha256=1b4e7ac9864e5bd1fd95aef0f98c733dfb12d8b5d155f54b4716165935401982
  Stored in directory: c:\users\samir\appdata\local\pip\cache\wheels\28\71\e4\38b5d81438105d0e3db5016cf2eea6fa796d89d96a04451d4d
  Building wheel for pathtools (setup.py) ... done
  Created wheel for pathtools: filename=pathtools-0.1.2-py3-none-any.whl size=8801 sha256=49e958d9d60e872e8ff60cfe6f5e5bced79b0f6cce67bf2dc5203b3c2f394208
  Stored in directory: c:\users\samir\appdata\local\pip\cache\wheels\4c\8e\7e\72fbc243e1aeecae64a96875432e70d4e92f3d2d18123be004
  Building wheel for urwid_readline (pyproject.toml) ... done
  Created wheel for urwid_readline: filename=urwid_readline-0.13-py3-none-any.whl size=7554 sha256=8136581f87d7d5a4ad479e6c0a95c62c2c142e6b92197cebd1e29e1a3c9406cd
  Stored in directory: c:\users\samir\appdata\local\pip\cache\wheels\80\10\cf\57c5430b5e5033995b239456813efa2587b486130a2688ab3d
  Building wheel for pyperclip (setup.py) ... done
  Created wheel for pyperclip: filename=pyperclip-1.8.2-py3-none-any.whl size=11137 sha256=37ab5a575353e7b897b4556866d52f97c4002f262633ed3453577e0d37be3fa6
  Stored in directory: c:\users\samir\appdata\local\pip\cache\wheels\7f\1a\65\84ff8c386bec21fca6d220ea1f5498a0367883a78dd5ba6122
Successfully built pudb antlr4-python3-runtime urwid pathtools urwid_readline pyperclip
Failed to build pycocotools
ERROR: Could not build wheels for pycocotools, which is required to install pyproject.toml-based projects
