### 2023-08-31 #5625 @just-eoghan
> @samirmartins Thank you for looking into this downloading issue and for your review!
>
> @AnnikaStein @ns-rse did you run into the same issue? @just-eoghan I'll try this myself, but please advise.

Hey, thanks for all your efforts on the review so far!

For the downloading issue, the reason for choosing Cityscapes is because the [standard benchmarks](https://paperswithcode.com/sota/unsupervised-domain-adaptation-on-cityscapes-1) for this task are based on that dataset.

This makes it easy to prove that the results from our software, which is much easier to run than the official implementations, correlate with the benchmarks, as can be seen in the [provenance logs](https://wandb.ai/eoghan/Lightning-UDA-detect)

The dataset is large and cumbersome, but no standard open-source dataset for benchmarking this task exists.

I didn't run into any issues downloading the dataset, except for creating an account and being approved.
