### 2023-09-11 #5625 @samirmartins
Dear colleagues,

I managed to download the files and test the code. Both the preprocessing script and the main code work as expected. However, the other questions I raised have not yet been addressed by the authors of the paper.

Sincerely,

Samir Martins
