### 2023-09-16 #5625 @ns-rse
Apologies, been on annual leave.

@galessiorob I had some challenges downloading the data at home (by virtue of a slow connection and lack of patience) but was able to download the test data at work in reasonable time (University with decent connection). I had no issues registering to do so.
