### 2023-10-24 #5625 @galessiorob
Hi everyone, and apologies for the very delayed response. Circling back on what needs to be done to close this paper:

- @AnnikaStein will you be able to complete this review in the upcoming month?
- @ns-rse thanks for the extra legwork to get the download working and test the software. @just-eoghan I had no issues but it did take some time, I know that this data set is needed for the benchmark, maybe if there are any smaller ones in the future it'd be worth swapping it. You could add a small note about this so that it's documented.
- @samirmartins thank you so much for your continuous support to get this reviewed thoroughly. Is the last question you still haven't gotten an answer to: _"All commits were made solely by @just-eoghan. Where are the contributions from the other authors of the work?"_ ?

@just-eoghan from the reviewer's feedback and the items that need to be addressed:
1. **Installation:** The paper does mention that "Lightning-UDA-Detect is structured for straightforward installation" but does not provide explicit details on how the installation proceeds. The paper also mentions that integrating with torchvision "reduces the dependencies of Lightning-UDA-Detect and removes the need to compile from source," however, specific installation steps or a link to documentation detailing the installation process are not mentioned. @just-eoghanyou could add a link to your instructions [here](https://github.com/just-eoghan/Lightning-UDA-detect/tree/master#environment--package-install) and a short explanation.

2. **Functionality:** The paper confirms the functional claims of the software by comparing the performance of their re-implementations with the officially reported performance of several UDA architectures, it also states that their models achieve or even exceed the official mAP@50 scores. @samirmartins @ns-rse are the review points that allude to functionality satisfied in your opinion with the current version?

3. **Automated tests:** Automated tests are not explicitly mentioned, however, the paper a detailed account of the performance evaluation conducted, including mAP@50 validation graphs and maximum mAP@50 scores. The training process and environment (e.g., the type of GPU used, the number of steps, and the seed) are described. If this manual test is meant to replace the expected automated tests, then I suggest adding an explanation as to why.

4. **Community guidelines:** The paper mentions that "Contributions to the package are warmly welcome; please open an issue to discuss new features." but the paper does not provide detailed community guidelines nor are they provided in the repo. Please add, here are [some examples](https://opensource.creativecommons.org/contributing-code/) and [tips.](https://opensource.com/life/16/3/contributor-guidelines-template-and-tips)
