### 2023-10-26 #5625 @just-eoghan
Hi all,

Thank you for your work! Looking forward to getting the last requirements fulfilled to close this paper.

@galessiorob I will remediate the points raised above in relation to installation, functionality, automated tests and community guidelines and revert when complete.

In relation to @samirmartins on _Where are the contributions from the other authors_ these authors are my PhD supervisors and contributed with guidance and mentorship for the project.
