### 2023-11-07 #5625 @ns-rse
@just-eoghan

> Dear @galessiorob,
>
> * I suggest that the authors create a pip installer (PyPI) to make the process more user-friendly. As it stands, the final user has to install some unnecessary libraries (such as black) to run the scripts.

If its of any use I've written about [ns-rse - Python Packaging](https://ns-rse.github.io/posts/python-packaging/) (mirrored on my [employers blog](https://rse.shef.ac.uk/blog/2023-09-18-python-packaging/)).
