### 2023-12-03 457edb3
Update README.md
Add detailed install instructions
