### 2023-12-03 840065c
Update requirements.txt
remove pycocotools, change to install step to facilitate windows installers
