### 2023-12-03 57f9244
Update README.md
add reasoning for testing format
