### 2023-12-11 #5625 @galessiorob
Thanks so much for[ adding these ](https://github.com/openjournals/joss-reviews/issues/5625#issuecomment-1837588443)@just-eoghan !

@samirmartins @ns-rse can you please check the above additions and review the last items on your checklist?

@AnnikaStein given we have not received any updates for your review I'll remove you from this paper so we can go ahead and publish as soon as the other editors' checklists are complete. I hope we can work together in a future paper.
