### 2023-12-18 #5625 @samirmartins
> Dear @galessiorob,
>
> The points that the authors have not yet considered/answered in my review are:
>
>     * I suggest that the authors create a pip installer (PyPI) to make the process more user-friendly. As it stands, the final user has to install some unnecessary libraries (such as black) to run the scripts.
>
>     * I encountered an error during the installation of requirements.txt on Windows 11 (resolved by installing pycocotools-windows instead of the pycocotools from requirements). How can we resolve this OS-dependent issue? The error description can be found at the end of this message.
>
>     * According to the JOSS Reviewer Guidelines, the article should have documentation, tests, continuous integration, and a license. Some highlights: i) I did not find any tests in the repository; ii) The documentation is very concise and should be improved (currently, it is limited to README.md). As a suggestion, consider using mkdocs or a similar tool to build it.
>
>     * All commits were made solely by @just-eoghan. Where are the contributions from the other authors of the work?
>
>     * There are no clear guidelines on how to contribute to the software/report issues/seek support.

Dear @galessiorob @just-eoghan ,

Regarding this review, I have the following comments:

"- I suggest that the authors create a pip installer (PyPI) to make the process more user-friendly. As it stands, the final user has to install some unnecessary libraries (such as black) to run the scripts."

Can't check this item, as it has not been addressed yet.

"I encountered an error during the installation of requirements.txt on Windows 11 (resolved by installing pycocotools-windows instead of the pycocotools from requirements). How can we resolve this OS-dependent issue? The error description can be found at the end of this message."

After this iteration, the authors have clarified this in README.md. I'll consider it done.

"According to the JOSS Reviewer Guidelines, the article should have documentation, tests, continuous integration, and a license. Some highlights: i) I did not find any tests in the repository; ii) The documentation is very concise and should be improved (currently, it is limited to README.md). As a suggestion, consider using mkdocs or a similar tool to build it.
"

i) I still could not find any tests in the repository (the authors should consider using pytest for this); ii) The documentation is still very concise, limited to README.md. Can't check this item.

"All commits were made solely by @just-eoghan. Where are the contributions from the other authors of the work?"

This was clarified by the authors. Consider it done.

"There are no clear guidelines on how to contribute to the software/report issues/seek support."

They are now present in the current version. Consider it done.

There are also some reference issues, detailed above by @ns-rse. Can't check this item either.

I've updated the checklist, but there are still some issues that need to be solved before publication.

Best regards,

Samir
