### 2024-01-08 #5625 @galessiorob
Happy New Year!

Thanks for your comments @samirmartins. @just-eoghan could you please review and address where applicable, thanks!
