### 2024-06-03 #5625 @galessiorob
Hey @crvernon, there's no email for @just-eoghan on the JOSS db and I couldn't find it on their Linkedin. How should we proceed?
